feat(warehouse-price): add file existence check and clear field functionality
Add a check before writing warehouse standards to file to confirm if a file with
the same name already exists. If it does, prompt the user to confirm whether to
overwrite the file. Also, implement a clear field function to reset all input
fields and display a success message to the user.

+++ controllers/warehouse_price.py
@@ -19,7 +19,8 @@ class warehouse_price():
     def __init__(self, main_window):
         self.main_window = main_window
         self.check_config_folder()
-        self.main_window.wright_yaml.currentIndexChanged.connect(self.write_warehouse_stander)
+        self.main_window.wright_yaml.clicked.connect(self.write_warehouse_stander)
+        self.main_window.clean_cfs_charge.clicked.connect(self.clear_field)
 
     
 
@@ -63,6 +64,51 @@ class warehouse_price():
             "cfs_ows_charge": cfs_ows_charge,
             "cfs_Insurance_charge": cfs_Insurance_charge,
         }
-        #写入文件
-        with open(f"conf/warehouse/{stander_name}.yaml", "w", encoding="utf-8") as f:
-            yaml.dump(cfs_stander, f, allow_unicode=True, default_flow_style=False)
+        #检查warehouse文件夹中是否已经由对应的名称
+        if os.path.exists(f"conf/warehouse/{stander_name}.yaml"):
+            #弹出框询问是否覆盖
+            if QMessageBox.question(
+                self.main_window,
+                "提示",
+                "文件已存在，是否覆盖？",
+                QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No,
+            ) == QMessageBox.StandardButton.Yes:
+                with open(f"conf/warehouse/{stander_name}.yaml", "w", encoding="utf-8") as f:
+                    yaml.dump(cfs_stander, f, allow_unicode=True, default_flow_style=False)
+        else:
+            #写入文件
+            with open(f"conf/warehouse/{stander_name}.yaml", "w", encoding="utf-8") as f:
+                yaml.dump(cfs_stander, f, allow_unicode=True, default_flow_style=False)
+
+    #清除字段
+    def clear_field(self):
+        self.main_window.cfs_name.clear()
+        self.main_window.cfs_bl_charge.clear()
+        self.main_window.cfs_yg_pkgs_charge.clear()
+        self.main_window.cfs_yg_weight_charge.clear()
+        self.main_window.cfs_yg_cmb_charge.clear()
+        self.main_window.cfs_yg_mini_charge.clear()
+        self.main_window.cfs_ys_pkgs_charge.clear()
+        self.main_window.cfs_ys_weight_charge.clear()
+        self.main_window.cfs_ys_cbm_charge.clear()
+        self.main_window.cfs_ys_mini_charge.clear()
+        self.main_window.cfs_van_charge.clear()
+        self.main_window.cfs_ows_charge.clear()
+        self.main_window.cfs_Insurance_charge.clear()
+        #弹框
+        QMessageBox.information(self.main_window, "提示", "清除成功")
+
+    
+    #读取conf/warehouse 文件并呈现在QlistView中对象名为stander_list
+    def read_warehouse_stander(self):
+        # 创建表格模型并填充数据
+        model = QStandardItemModel()
+        header_labels = ["名称"]
+        model.setHorizontalHeaderLabels(header_labels)
+        for file in os.listdir("conf/warehouse"):
+            if file.endswith(".yaml"):
+                item = QStandardItem(file.replace(".yaml", ""))
+                model.appendRow([item])
+        self.main_window.stander_list.setModel(model)
+
+
